pyproject.toml
[project]
name = "real-estate-rag-chatbot"
version = "0.1.0"
description = "Conversion-focused RAG chatbot for Al Badia Villas"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "pdfminer.six>=20221105",
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=2.7.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "openai>=1.40.0",
  "httpx>=0.27.0",
  "orjson>=3.10.7",
  "regex>=2024.5.10",
  "pytest>=8.3.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
